fix tests/lint, add crates to cargo workspace, fix CI
