+++ crates/stratus_macros/Cargo.toml
@@ -10,3 +10,6 @@ proc-macro = true
 syn = { version = "=2.0.104", features = ["full"] }
 quote = "=1.0.47"
 proc-macro2 = "=1.0.95"
+
+[lints]
+workspace = true
