+++ justfile_helpers
@@ -1,6 +1,6 @@
 _lint version="" fmt-flag="" clippy-flag="":
     cargo +nightly{{version}} fmt --all {{fmt-flag}}
-    cargo +nightly{{version}} clippy --all-targets -- {{clippy-flag}}
+    cargo +nightly{{version}} clippy --workspace --all-targets -- {{clippy-flag}}
 
 _doc version="":
     cargo +nightly{{version}} doc --no-deps
