+++ src/eth/follower/importer/mod.rs
@@ -270,11 +270,7 @@ mod tests {
     use hash_hasher::HashBuildHasher;
 
     use super::BlockchainClient;
-    use crate::eth::executor::ExecutionResult;
     use crate::eth::executor::State;
-    use crate::eth::executor::TransactionExecution;
-    use crate::eth::executor::TransactionExecutionInput;
-    use crate::eth::executor::TransactionExecutionResult;
     use crate::eth::executor::types::state::AccountChanges;
     use crate::eth::executor::types::state::Complete;
     use crate::eth::executor::types::state::CompleteValue;
@@ -295,9 +291,6 @@ mod tests {
     use crate::eth::types::Block;
     use crate::eth::types::BlockNumber;
     use crate::eth::types::PointInTime;
-    use crate::eth::types::Signature;
-    use crate::eth::types::TransactionInfo;
-    use crate::eth::types::TransactionInput;
     use crate::eth::types::UnixTime;
     use crate::eth::types::Wei;
 
@@ -309,23 +302,6 @@ mod tests {
                 bytecode: CompleteValue::Changed(account.bytecode),
             }
         }
-    }
-
-    /// Mines a block applying `changes` (mirrors the helper in `stratus_storage` tests).
-    fn mine_block(storage: &StratusStorage, state: State<Complete>) {
-        let header = storage.read_pending_block_header();
-        let evm_input = TransactionExecutionInput::create(&TransactionInput::default(), header);
-
-        let result = TransactionExecutionResult {
-            result: ExecutionResult::Success,
-            ..Default::default()
-        };
-
-        let tx = TransactionExecution::new(TransactionInfo::default(), Signature::default(), evm_input, result);
-        storage.save_execution(tx, state).expect("save execution");
-
-        let (block, block_changes) = storage.finish_pending_block();
-        storage.save_block(block.into(), block_changes).expect("save block");
     }
 
     /// Builds `ExecutionChanges` that set `address`'s balance to `balance` (nonce/bytecode untouched).
@@ -386,7 +362,7 @@ mod tests {
         let address = Address::new([0xCC; 20]);
 
         // Block 1: B.balance = 100. permanent storage is now at block 1.
-        mine_block(&storage, balance_changes(address, Wei::from(100u64)));
+        storage.mine_block_with_mock_execution(balance_changes(address, Wei::from(100u64)));
 
         // The fetcher post-processes block 3 while the importer is still at block 1 (fetcher ahead).
         // Block 3 changed B's nonce but left its balance untouched (balance entry is `None`).
@@ -400,7 +376,7 @@ mod tests {
 
         // Intervening block 2: B.balance = 200. This is the correct pre-state for block 3.
         // permanent storage is now at block 2.
-        mine_block(&storage, balance_changes(address, Wei::from(200u64)));
+        storage.mine_block_with_mock_execution(balance_changes(address, Wei::from(200u64)));
 
         // The importer imports block 3. `ReplicationWorker::import` must complete `changes_3`
         // (Incomplete) against perm at import time, when perm is caught up to block 2 (200).
